% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_collection.R
\name{declass_collection}
\alias{declass_collection}
\title{List collections.}
\usage{
declass_collection(return_type = "string")
}
\arguments{
\item{return_type}{The format of content displayed, in list or raw.}
}
\value{
The name of collections.
}
\description{
This functions lists all collections that're avaialble in the Declassification Engine API.
}
\examples{
declass_collection("list")
declass_collection("string")
}
