% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_collection_entity.R
\name{declass_collection_entity}
\alias{declass_collection_entity}
\title{Display entities for a collection.}
\usage{
declass_collection_entity(collection)
}
\arguments{
\item{collection}{The collection name.}
}
\value{
The entities for the collection.
}
\description{
This functions display all available entities for a specified collection.
}
\examples{
declass_collection_entity("clinton")
declass_collection_entity("kissinger")
declass_collection_entity("frus")
}
