% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_entity_data.R
\name{declass_entity_data}
\alias{declass_entity_data}
\title{Query data of an entity in a collection.}
\usage{
declass_entity_data(collection, entity, page_size = 25, page = 1,
  num_pages = 10)
}
\arguments{
\item{collection}{The collection name}

\item{entity}{The entity name}

\item{page_size}{The number of rows displayed per page. Default setting is 25 records per page.}

\item{page}{The starting page number to be queried. Default setting is page No.1.}

\item{num_pages}{The number of pages to be quried. Default setting is 10 pages.}
}
\value{
Fields of data for the specified entity in that collection, and urls.
}
\description{
This functions queries data of a specified entitiy in a specified collection.
}
\examples{
declass_entity_data("kissinger", "countries", page_size = 30, page = 1, num_pages = 5)
}
