% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_fields.R
\name{declass_fields}
\alias{declass_fields}
\title{Display document metadata fields.}
\usage{
declass_fields(return_type = "string")
}
\arguments{
\item{return_type}{The format of content displayed, in list or character.}
}
\value{
The metadata fields for all collections of data.
}
\description{
This function displays the document metadata fields for all collections.
}
\examples{
declass_fields("list")
declass_fields("string")
}
