% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_query_id_data.R
\name{declass_query_id_data}
\alias{declass_query_id_data}
\title{Query document by ID.}
\usage{
declass_query_id_data(id_list)
}
\arguments{
\item{id_list}{A list of IDs}
}
\value{
Content of the entire document
}
\description{
This function queries a specific document or documents by user-inputed ID. It is intended for users who know what they want and have specific IDs in mind.
}
\examples{
declass_query_id_data("Clinton-74931")
declass_query_id_data("1977LONDON17830,1975MOSCOW15238,P790162-1195")
}
