% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_query_random_ids.R
\name{declass_query_random_ids}
\alias{declass_query_random_ids}
\title{Return random document IDs with a limit.}
\usage{
declass_query_random_ids(limit = 10)
}
\arguments{
\item{limit}{The number of document IDs to be randomly sampled. Default setting is 10 ids. Maximum is set to be 1024, so as not to crash the API server.}
}
\value{
Document IDs.
}
\description{
This function returns randomly sampled document IDs. It is intended for users who don't know what to look for, and it lets them explore and view the documents.
}
\examples{
declass_query_random_ids(32)
declass_query_random_ids(1100)
}
