% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declass_query_random_ids_data.R
\name{declass_query_random_ids_data}
\alias{declass_query_random_ids_data}
\title{Query documents based on randomly sampled IDs.}
\usage{
declass_query_random_ids_data(limit)
}
\arguments{
\item{limit}{The number of randomply sampled IDs.}
}
\value{
Content of the documents
}
\description{
This function queries documents by randomly sampled IDs.
}
\examples{
declass_query_random_ids_data(3)
}
